requirements.txt
Django==1.3.1
south>=0.7
ella>=3.0.0
django-tagging